pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "freeagent"
version = "0.1.0"
description = "FreeAgent — a free, open-source terminal coding agent, Claude Code-style, powered by z.ai's free GLM APIs."
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Vasily Sahrai" }]
requires-python = ">=3.9"
keywords = ["ai", "agent", "cli", "coding-assistant", "z.ai", "glm", "open-source"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Topic :: Software Development",
]
dependencies = [
  "openai>=1.40.0",
  "rich>=13.7.0",
  "prompt_toolkit>=3.0.47",
  "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/vasilysahrai/freeagent"
Repository = "https://github.com/vasilysahrai/freeagent"
Issues = "https://github.com/vasilysahrai/freeagent/issues"

[project.scripts]
freeagent = "freeagent.cli:main"

[tool.setuptools.packages.find]
include = ["freeagent*"]
